fix octal integer handling
In the previous version we use to represent an octal number with a
0-prefix. Since python3 has migrated to 0o-prefix rapresentation,
we decided to switch to our simple representation:
the octal number should be a simple 3 digits number.

In any case 0-prefix and 0o-prefix are supported as well.

To convert the dictionary that handle the configuration from json and
back to it, now format/unformat and unformat functions are used